separating models from database configuration, creating folders/files for routers, adding aws config file for when that's figured out
