+++ app.js
@@ -6,7 +6,7 @@ const port = 3001
 app.use(express.json());
 
 app.listen(port, () => {
-  console.log("Server is running on port 3001");
+  console.log("Server is running on port " + port);
 });
 
 module.exports = app;
