+++ routes/images.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const PatientImage = require('../models/PatientImage');
+const ImageProcessing = require('../models/ImageProcessing');
+
+// IMAGE UPLOAD PIPELINE GOES HERE ?? idk for sure lol
+
+module.exports = router;
